feat(push-regex): add a parameter for the component name
The push-regex command is great, but because it's very generic the commit message and Pull Request title are very hard to use.
This PR introduce a new parameter: `--component-name`, to specify an optional name for the "component" that is updated.
This name will be used in the commit message and Pull Request title to have nicer and more user-friendly messages.